+++ database/migrations/2025_11_19_054330_create_order_items_table.php
@@ -8,7 +8,8 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::create('order_items', function (Blueprint $table) {
+        if (!Schema::hasTable('order_items')) {
+            Schema::create('order_items', function (Blueprint $table) {
             $table->id();
             $table->foreignId('order_id')->constrained()->onDelete('cascade');
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
@@ -32,6 +33,7 @@ return new class extends Migration
 
             $table->timestamps();
         });
+        }
     }
 
     public function down()
